ABC081-090 A,B

+++ utils.cpp
@@ -126,7 +126,7 @@ ll div(ll a, ll b, ll m) { return (a * modpow(b, m - 2, m)) % m; }
 // 平方数判別
 bool is_square(ll n)
 {
-  for (int i = 1; i * i <= n; i++)
+  for (ll i = 1; i * i <= n; i++)
     if (i * i == n)
       return true;
   return false;
